Rename glWrappers as ComputeShader
